Add nucleotide solution
Add primitive solution for histogram that iterates the strand anew for every
nucleotide. This is commented out
Add solution that iterates the strand while simultaneously incrementing
the histogram map. Also this does some error handling if it accounters
a nucleotide that is unexpected i.e. not in the histogram map.
Add a test for the error handling